requirements.txt
Django==1.7.7
django-braces==1.4.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
easy-thumbnails==2.2
Pillow==2.8.1
python-social-auth==0.2.3
python3-openid==3.0.5
selenium==2.45.0